Refactor updateSelectedBBox
